requirements.txt
feedparser
apscheduler
python-dateutil
kafka-python>=1.3.3
